pyproject.toml
[tool.poetry]
name = "coreason-veritas"
version = "0.4.0"
description = "coreason_veritas is the non-negotiable governance layer of the CoReason platform (Prosperity Public License 3.0.0)"
authors = ["Gowtham A Rao <gowtham.rao@coreason.ai>"]
license = "Proprietary"
readme = "README.md"
packages = [{include = "coreason_veritas", from = "src"}]
classifiers = [
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]

[tool.poetry.urls]
Homepage = "https://github.com/CoReason-AI/coreason_veritas"
Repository = "https://github.com/CoReason-AI/coreason_veritas"
Documentation = "https://github.com/CoReason-AI/coreason_veritas"

[tool.poetry.dependencies]
python = ">=3.12, <3.15"
loguru = "^0.7.2"
cryptography = "^46.0.3"
opentelemetry-api = "^1.39.1"
pydantic = "^2.12.5"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
httpx = "^0.28.1"
opentelemetry-exporter-otlp = "^1.39.1"
opentelemetry-instrumentation-fastapi = "^0.60b1"
opentelemetry-sdk = "^1.39.1"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2.2"
ruff = "^0.4.8"
pre-commit = "^3.7.1"
pytest-cov = "^5.0.0"
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.26"
opentelemetry-sdk = "^1.39.1"
pytest-asyncio = "^1.3.0"
twine = "^6.2.0"
hypothesis = "^6.148.8"
respx = "^0.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["coreason_veritas"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=100"
testpaths = ["tests"]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
omit = ["tests/*"]
